refactor: features/product 배럴 패턴 적용
