+++ src/components/features/product/components/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductCard } from './ProductCard';
+export { default as ProductRowCard } from './ProductRowCard';
