+++ src/components/features/product/index.ts
@@ -1,3 +1,3 @@
-export { default as ProductCard } from './product-card/ProductCard';
-export { default as ProductRowCard } from './product-card/ProductRowCard';
+export * from './api';
+export * from './components';
 export type * from './type';
